+++ negotiation/src/main/java/it/eng/negotiation/exception/ContractNegotiationAPIExceptionAdvice.java
@@ -17,4 +17,10 @@ public class ContractNegotiationAPIExceptionAdvice extends ResponseEntityExcepti
     protected ResponseEntity<Object> handleContractNegotiationAPIException(ContractNegotiationAPIException ex, WebRequest request) {
     	return new ResponseEntity<>(GenericApiResponse.error(ex.getLocalizedMessage(), HttpStatus.BAD_REQUEST.value()), HttpStatus.BAD_REQUEST);
     }
+    
+    @ExceptionHandler(value = {ContractNegotiationInvalidStateException.class})
+    protected ResponseEntity<Object> handleContractNegotiationAPIException(ContractNegotiationInvalidStateException ex, WebRequest request) {
+    	return new ResponseEntity<>(GenericApiResponse.error(ex.getLocalizedMessage(), HttpStatus.BAD_REQUEST.value()), HttpStatus.BAD_REQUEST);
+    }
+    
 }